+++ sign_up.go
@@ -5,19 +5,22 @@ import (
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/cognitoidentityprovider"
+	"github.com/aws/aws-sdk-go-v2/service/cognitoidentityprovider/types"
 	"github.com/gofor-little/xerror"
 )
 
 // SignUp signs a new user up.
-func SignUp(ctx context.Context, emailAddress string, password string, passwordConfirmation string) (string, error) {
-	if password != passwordConfirmation {
-		return "", xerror.New("password and password confirmation missmatch")
-	}
-
+func SignUp(ctx context.Context, emailAddress string, password string) (string, error) {
 	output, err := CognitoClient.SignUp(ctx, &cognitoidentityprovider.SignUpInput{
-		ClientId: aws.String(CognitoUserPoolClientID),
+		ClientId: aws.String(CognitoClientID),
 		Password: aws.String(password),
 		Username: aws.String(emailAddress),
+		UserAttributes: []types.AttributeType{
+			{
+				Name:  aws.String("email"),
+				Value: aws.String(emailAddress),
+			},
+		},
 	})
 	if err != nil {
 		return "", xerror.Wrap("failed to sign up user", err)
@@ -29,7 +32,7 @@ func SignUp(ctx context.Context, emailAddress string, password string, passwordC
 // SignUpConfirm confirms a newly signed up user with the confirmation code they received.
 func SignUpConfirm(ctx context.Context, emailAddress string, confirmationCode string) error {
 	_, err := CognitoClient.ConfirmSignUp(ctx, &cognitoidentityprovider.ConfirmSignUpInput{
-		ClientId:         aws.String(CognitoUserPoolClientID),
+		ClientId:         aws.String(CognitoClientID),
 		ConfirmationCode: aws.String(confirmationCode),
 		Username:         aws.String(emailAddress),
 	})